+++ VacationMachineTest/PerformerEmployeeTest.cs
@@ -3,23 +3,23 @@ using NUnit.Framework;
 using VacationMachine;
 using VacationMachine.Business;
 
-namespace VacationMachineTest.Business
+namespace VacationMachineTest
 {
-    public class RegularEmployeeTest
+    public class PerformerEmployeeTest
     {
-        private const int EMPLOYEE_ID = 1;
-
         private IMessageBus _messageBus;
         private IEmailSender _emailSender;
-        private RegularEmployee _sut;
+        private IEscalationManager _escalationManager;
+        private PerformerEmployee _sut;
 
         [SetUp]
         public void Initialize()
         {
             _messageBus = Substitute.For<IMessageBus>();
             _emailSender = Substitute.For<IEmailSender>();
+            _escalationManager = Substitute.For<IEscalationManager>();
 
-            _sut = new RegularEmployee(_messageBus, _emailSender);
+            _sut = new PerformerEmployee(_messageBus, _emailSender, _escalationManager);
         }
 
         [Test]
@@ -34,9 +34,19 @@ namespace VacationMachineTest.Business
         }
 
         [Test]
-        [TestCase(Configuration.MAX_DAYS + 1)]
+        [TestCaseSource(typeof(TestCaseHelper), nameof(TestCaseHelper.GetDaysFromTo), new object[] { Configuration.MAX_DAYS + 1, Configuration.MAX_DAYS_FOR_PERFORMERS })]
+        public void RequestPaidDaysOff_WhenDaysAboveMaxDaysAndBelowMaxDaysForPerformesRequested_ThenManual(int days)
+        {
+            var expectedResult = typeof(ManualRequestResult);
+
+            var actualResult = _sut.RequestPaidDaysOff(days).GetType();
+
+            Assert.AreEqual(expectedResult, actualResult);
+        }
+
+        [Test]
         [TestCase(Configuration.MAX_DAYS_FOR_PERFORMERS + 1)]
-        public void RequestPaidDaysOff_WhenDaysAboveMaxDaysRequested_ThenDenied(int days)
+        public void RequestPaidDaysOff_WhenDaysAboveMaxDaysForPerformesRequested_ThenDenied(int days)
         {
             var expectedResult = typeof(DeniedRequestResult);
