+++ VacationMachineTest/RegularEmployeeTest.cs
@@ -3,25 +3,21 @@ using NUnit.Framework;
 using VacationMachine;
 using VacationMachine.Business;
 
-namespace VacationMachineTest.Business
+namespace VacationMachineTest
 {
-    public class PerformerEmployeeTest
+    public class RegularEmployeeTest
     {
-        private const int EMPLOYEE_ID = 1;
-
         private IMessageBus _messageBus;
         private IEmailSender _emailSender;
-        private IEscalationManager _escalationManager;
-        private PerformerEmployee _sut;
+        private RegularEmployee _sut;
 
         [SetUp]
         public void Initialize()
         {
             _messageBus = Substitute.For<IMessageBus>();
             _emailSender = Substitute.For<IEmailSender>();
-            _escalationManager = Substitute.For<IEscalationManager>();
 
-            _sut = new PerformerEmployee(_messageBus, _emailSender, _escalationManager);
+            _sut = new RegularEmployee(_messageBus, _emailSender);
         }
 
         [Test]
@@ -36,19 +32,9 @@ namespace VacationMachineTest.Business
         }
 
         [Test]
-        [TestCaseSource(typeof(TestCaseHelper), nameof(TestCaseHelper.GetDaysFromTo), new object[] { Configuration.MAX_DAYS + 1, Configuration.MAX_DAYS_FOR_PERFORMERS })]
-        public void RequestPaidDaysOff_WhenDaysAboveMaxDaysAndBelowMaxDaysForPerformesRequested_ThenManual(int days)
-        {
-            var expectedResult = typeof(ManualRequestResult);
-
-            var actualResult = _sut.RequestPaidDaysOff(days).GetType();
-
-            Assert.AreEqual(expectedResult, actualResult);
-        }
-
-        [Test]
+        [TestCase(Configuration.MAX_DAYS + 1)]
         [TestCase(Configuration.MAX_DAYS_FOR_PERFORMERS + 1)]
-        public void RequestPaidDaysOff_WhenDaysAboveMaxDaysForPerformesRequested_ThenDenied(int days)
+        public void RequestPaidDaysOff_WhenDaysAboveMaxDaysRequested_ThenDenied(int days)
         {
             var expectedResult = typeof(DeniedRequestResult);
